Moved ios_system setup to AppDelegate
